+++ src/main/java/jungjin/board/repository/BoardMasterRepository.java
@@ -4,5 +4,5 @@ import jungjin.board.domain.BoardMaster;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface BoardMasterRepository extends JpaRepository<BoardMaster, Long> {
+public interface BoardMasterRepository extends JpaRepository<BoardMaster, Integer> {
 }
